♻️ refactor: 서드파티라이브러리 파일 이동 - infra/sdk -> common/libs - 경로 변경으로 인한 추가적인 경로 수정
