+++ src/pages/_app.tsx
@@ -11,8 +11,8 @@ import { Layout } from "@/common/components/Layout";
 import { SignUpModal, SignUpModalProvider } from "@/common/components/Modal";
 import { ToastContainer, ToastProvider } from "@/common/components/Toast";
 import { OverlayProvider, RouteTrackingProvider, useAnalytics } from "@/common/hooks";
+import { GoogleTagManagerScript, GTagScript } from "@/common/libs";
 import { TagCategoryProvider } from "@/components/tags";
-import { GoogleTagManagerScript, GTagScript } from "@/infra/sdk";
 import type { DefaultPageProps } from "@/types";
 
 if (process.env.NEXT_PUBLIC_API_MOCKING === "enabled") {
